requirements.txt
Flask==3.0.3
pytrends==4.9.2
google-generativeai>=0.3.0
requests>=2.31.0
libsql-client>=0.1.0
